Trait pallet_example_offchain_worker::Trait

pub trait Trait: CreateSignedTransaction<Call<Self>> {
    type AuthorityId: AppCrypto<Self::Public, Self::Signature>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type Call: From<Call<Self>>;
    type GracePeriod: Get<Self::BlockNumber>;
    type UnsignedInterval: Get<Self::BlockNumber>;
    type UnsignedPriority: Get<TransactionPriority>;
}

This pallet’s configuration trait

Associated Types

type AuthorityId: AppCrypto<Self::Public, Self::Signature>

The identifier type for an offchain worker.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type Call: From<Call<Self>>

The overarching dispatch call type.

type GracePeriod: Get<Self::BlockNumber>

A grace period after we send transaction.

To avoid sending too many transactions, we only attempt to send one every GRACE_PERIOD blocks. We use Local Storage to coordinate sending between distinct runs of this offchain worker.

type UnsignedInterval: Get<Self::BlockNumber>

Number of blocks of cooldown after unsigned transaction is included.

This ensures that we only accept unsigned transactions once, every UnsignedInterval blocks.

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.

Loading content...